refactor : 검색 페이지 리팩토링 - 헤더 컴포넌트 분리 및 스타일 통일 - 공용 스타일 파일 분리 - 실시간 검색 fetch함수 이름 변경
