+++ src/components/Explore/Search/RealTimeSearch/RealTimeSearchResultPanel.tsx
@@ -1,15 +1,15 @@
 /** @jsxImportSource @emotion/react */
 
 import { Place } from "../../../../types/types";
-import { SearchList } from "../RecentSearch/RecentSearchesPanel";
-import { useDeferredValue, useEffect, useState, useTransition } from "react";
+import { useEffect, useState, useTransition } from "react";
 import { getRequest } from "../../../../api/utils/getRequest";
 import { isLastCharKoreanConsonantOrVowel } from "../../../../utils/checkLastChar";
 import useCurrentLocation from "../../../../hooks/map/useCurrentLocation";
-import useDeboucing from "../../../../hooks/useDeboucing";
+import { SearchList } from "../styles/Search.styles";
 
 import RealTimeSearchItem from "./RealTimeSearchItem";
 import NoSearchData from "../NoSearchData";
+import SearchPannelHeader from "../SearchPannelHeader";
 
 interface RealtimeSearchResultPanelProps {
   searchQuery: string;
@@ -19,42 +19,37 @@ const RealtimeSearchResultPanel = ({
   searchQuery,
 }: RealtimeSearchResultPanelProps) => {
   const [realTimeSearchResult, setRealTimeSearchResult] = useState<Place[]>([]);
-
   const [isPending, startTransition] = useTransition();
-
   const { currentLocation } = useCurrentLocation();
-
-  const deferredSearchQuery = useDeferredValue(searchQuery);
-  const { debouncedInput: debouncedSearchQuery } = useDeboucing(
-    deferredSearchQuery,
-    300
-  );
 
   useEffect(() => {
     startTransition(() => {
-      const getSearchResult = async () => {
-        if (!currentLocation || !debouncedSearchQuery) return;
-        if (isLastCharKoreanConsonantOrVowel(debouncedSearchQuery)) return;
+      const fetchRealTimeSearchResult = async () => {
+        if (!currentLocation || !searchQuery) return;
+        if (isLastCharKoreanConsonantOrVowel(searchQuery)) return;
 
         const x = currentLocation?.lng!;
         const y = currentLocation?.lat!;
 
         const searchResult = await getRequest({
-          url: `v1/places/search?query=${debouncedSearchQuery}&x=${x}&y=${y}`,
+          url: `v1/places/search?query=${searchQuery}&x=${x}&y=${y}`,
         });
         setRealTimeSearchResult(searchResult);
       };
-      getSearchResult();
+      fetchRealTimeSearchResult();
     });
-  }, [debouncedSearchQuery, currentLocation, startTransition]);
+  }, [searchQuery, currentLocation, startTransition]);
 
   return (
-    <SearchList>
-      {realTimeSearchResult.map((place: Place) => (
-        <RealTimeSearchItem key={place.id} place={place} />
-      ))}
-      {realTimeSearchResult.length === 0 && !isPending && <NoSearchData />}
-    </SearchList>
+    <>
+      <SearchPannelHeader title="검색 결과" />
+      <SearchList>
+        {realTimeSearchResult.map((place: Place) => (
+          <RealTimeSearchItem key={place.id} place={place} />
+        ))}
+        {realTimeSearchResult.length === 0 && !isPending && <NoSearchData />}
+      </SearchList>
+    </>
   );
 };
